Handle rbenv Ruby version manager in .bashrc
(rbenv)[https://github.com/rbenv/rbenv] is a Ruby version manager.

It inserts "shims" into the beginning of PATH to allow custom Ruby
versions to be used.

It needs `eval $(rbenv init -)` to be run. However, since this adds the
shims directory to the beginning of PATH, shells started within shells
all with the same .bashrc will have the shims directory repeated in
PATH. The sed expression removes repeats of the shims directory.