This fixes a problem where tmux exits immediately
This seems to be due to the referenced executable (the one the ref to
which I wrapped in the tmux conf conditional) not existing.

I tried process of elimination, and this works on Linux.

I need to see if this works on Mac.

Specifically, I need to quit out of tmux, deploy this branch (on a Mac
with the local bash in place), launch a tmux session, and run `echo
$BASH_VERSION`. If it is the overridden version, then this config works.