Add aliases for aws cli via Docker
If anything goes wrong with a local Python interpreter, or maybe a
Homebrew package, invoking aws can fail. For example, I just got:

```
$ aws lambda help
-bash: /Users/dougnaphas/Library/Python/3.7/bin/aws:
/usr/local/opt/python/bin/python3.7: bad interpreter: No such file
 or directory
```

This commit adds aliases based on
https://docs.aws.amazon.com/cli/latest/userguide/install-cliv2-docker.html
for running the AWS CLI tool with Docker, so that local files are less
likely to cause a problem.

awsman adds -it, because without it output did not appear correctly in
the pager.