Match bash_completion string from nvm
gh-7

The install script looks for the bash_completion string without {}
around the variable name.

Without this change, the nvm install script will append the
bash_completion string to .bashrc, thus making .bashrc stateful, which
is what I am trying to avoid with this project.

Here is the string the install script looks for:

https://github.com/nvm-sh/nvm/blob/1ffa41886a2311851766f17f2a0e5aa9d460be3d/install.sh#L320