Move brew in PATH priority below shims
Before this change:
$ echo ${PATH} | awk -v RS=':' {print}
/opt/homebrew/bin
/opt/homebrew/sbin
/Users/douglasnaphas/.nvm/versions/node/v16.13.0/bin
/usr/local/bin
/usr/bin
/bin
/usr/sbin
/sbin
/opt/homebrew/sbin

After:
$ echo $PATH | awk -v RS=':' {print}
/Users/douglasnaphas/.nvm/versions/node/v12.22.7/bin
/opt/homebrew/bin
/opt/homebrew/sbin
/usr/local/bin
/usr/bin
/bin
/usr/sbin
/sbin
/opt/homebrew/sbin

I am making this change so that things that shim or mess with the PATH
to make sure I'm using a specified version of Python, Node, or Ruby, for
example, are not overridden by a conflicting name owned by Homebrew.

Specifically, I need this change to be able to set a version of Node
that is different from the one installed via Homebrew.