Don't put brew before shims in PATH
gh-9

Not sure why, but this seems to be consistently putting the brew
directories last in the PATH, so I don't currently have a need to
further pursue ensuring brew is last in the PATH.

The part where I'm changing the means of testing whether this is an
interactive shell is from
https://tldp.org/LDP/abs/html/intandnonint.html#IITEST, and is done so
that I can test the function set in .bashrc. Without changing to
checking PS1 to determine interactivity, I was unable to test .bashrc,
because when I ran, for example, dedup_awk.test.sh, .bashrc would do
nothing because ./dedup_awk.test.sh resulted in .bashrc being run in a
non-interactive shell. Now that I'm checking PS1 to determine
interactivity, I can set PS1 when I run the test script, so .bashrc
thinks it's running interactively.